sr/model changes
