+++ apps/gotask_api/src/domain/model/asset/interface/access.ts
@@ -1,0 +1,8 @@
+import { IAsset } from "./asset";
+
+export interface IAccessCardAsset extends IAsset {
+  accessCardNo?: string;
+  accessCardNo2?: string;
+  personalId?: string;
+  issuedOn?: string;
+}
