+++ apps/gotask_api/src/domain/model/asset/interface/fingerPrint.ts
@@ -1,0 +1,10 @@
+import { IAsset } from "./asset";
+
+export interface IFingerprintScannerAsset extends IAsset {
+  location?: string;
+  connectivity?: string;
+  capacity?: string;
+  authenticationModes?: string;
+  display?: string;
+  cloudAndAppBased?: boolean;
+}
