+++ apps/gotask_api/src/domain/model/asset/interface/laptop.ts
@@ -1,0 +1,10 @@
+import { IAsset } from "./asset";
+
+export interface ILaptopAsset extends IAsset {
+  antivirus?: boolean;
+  recoveryKey?: string;
+  isEncrypted?: boolean;
+  lastServicedDate?: Date;
+  commentService?: string;
+  erk?: string;
+}
