+++ apps/gotask_api/src/domain/model/asset/interface/mobile.ts
@@ -1,0 +1,13 @@
+import { IAsset } from "./asset";
+
+export interface IMobileAsset extends IAsset {
+  imeiNumber?: string;
+  screenSize?: string;
+  batteryCapacity?: string;
+  cameraSpecs?: string;
+  simType?: string;
+  is5GSupported?: boolean;
+  insuranceProvider?: string;
+  insurancePolicyNumber?: string;
+  insuranceExpiry?: Date;
+}
