+++ apps/gotask_api/src/domain/model/asset/interface/printer.ts
@@ -1,0 +1,10 @@
+import { IAsset } from "./asset";
+
+export interface IPrinterAsset extends IAsset {
+  location?: string;
+  connectivity?: string;
+  printerType?: string;
+  specialFeatures?: string;
+  printerOutputType?: string;
+  supportedPaperSizes?: string;
+}
